Functional RS-232 RX bit-bang reader
Can now read bits from the serial port and flickers the first segment
display from 1 to 0 to confirm that a connection has been established.
An electrical connection error has misplaced the RX line on the TX line,
and the other control lines have been shifted down the length of the
chip as well. The code compensates for this.