Update dispSeg() in segment.c
Found and corrected a problem in the switch/case block, which prevented
it from properly processing alpha characters. The emulator is now known
to function and process commands correctly, at least to the extent that
it was tested for this commit. The test code sent immediate data to a
port that made the first segment display show "F".